feat: Initial commit of moving stack analysis for npm package manager to new exhort java api
